Struct emath::align::Align2

pub struct Align2(pub [Align; 2]);

Two-dimension alignment, e.g. Align2::LEFT_TOP.

Tuple Fields
0: [Align; 2]
Implementations

impl Align2

pub const LEFT_BOTTOM: Align2 = _

pub const LEFT_CENTER: Align2 = _

pub const LEFT_TOP: Align2 = _

pub const CENTER_BOTTOM: Align2 = _

pub const CENTER_CENTER: Align2 = _

pub const CENTER_TOP: Align2 = _

pub const RIGHT_BOTTOM: Align2 = _

pub const RIGHT_CENTER: Align2 = _

pub const RIGHT_TOP: Align2 = _

impl Align2

pub fn x(self) -> Align

Returns an alignment by the X (horizontal) axis

pub fn y(self) -> Align

Returns an alignment by the Y (vertical) axis

pub fn to_sign(self) -> Vec2

-1, 0, or +1 for each axis

pub fn anchor_rect(self, rect: Rect) -> Rect

Used e.g. to anchor a piece of text to a part of the rectangle. Give a position within the rect, specified by the aligns

pub fn align_size_within_rect(self, size: Vec2, frame: Rect) -> Rect

e.g. center a size within a given frame

pub fn pos_in_rect(self, frame: &Rect) -> Pos2

Returns the point on the rect’s frame or in the center of a rect according to the alignments of this object.

(*)-----------+------(*)------+-----------(*)--> X
 |            |               |            |
 |  Min, Min  |  Center, Min  |  Max, Min  |
 |            |               |            |
 +------------+---------------+------------+
 |            |               |            |
(*)Min, Center|Center(*)Center|Max, Center(*)
 |            |               |            |
 +------------+---------------+------------+
 |            |               |            |
 |  Min, Max  | Center, Max   |  Max, Max  |
 |            |               |            |
(*)-----------+------(*)------+-----------(*)
 |
 Y