emath
Struct Rangef

#[repr(C)]
pub struct Rangef {
    pub min: f32,
    pub max: f32,
}

Inclusive range of floats, i.e. min..=max, but more ergonomic than RangeInclusive.

Fields
min: f32max: f32
Implementations

impl Rangef

pub const EVERYTHING: Self = _

Infinite range that contains everything, from -∞ to +∞, inclusive.

pub const NOTHING: Self = _

The inverse of Self::EVERYTHING: stretches from positive infinity to negative infinity. Contains nothing.

pub const NAN: Self = _

An invalid Rangef filled with f32::NAN.

pub fn new(min: f32, max: f32) -> Self

pub fn point(min_and_max: f32) -> Self

pub fn span(self) -> f32

The length of the range, i.e. max - min.

pub fn center(self) -> f32

The center of the range

pub fn contains(self, x: f32) -> bool

pub fn clamp(self, x: f32) -> f32

Equivalent to x.clamp(min, max)

pub fn as_positive(self) -> Self

Flip min and max if needed, so that min <= max after.

pub fn shrink(self, amnt: f32) -> Self

Shrink by this much on each side, keeping the center

pub fn expand(self, amnt: f32) -> Self

Expand by this much on each side, keeping the center

pub fn flip(self) -> Self

Flip the min and the max

pub fn intersection(self, other: Self) -> Self

The overlap of two ranges, i.e. the range that is contained by both.

If the ranges do not overlap, returns a range with span() < 0.0.

assert_eq!(Rangef::new(0.0, 10.0).intersection(Rangef::new(5.0, 15.0)), Rangef::new(5.0, 10.0));
assert_eq!(Rangef::new(0.0, 10.0).intersection(Rangef::new(10.0, 20.0)), Rangef::new(10.0, 10.0));
assert!(Rangef::new(0.0, 10.0).intersection(Rangef::new(20.0, 30.0)).span() < 0.0);

pub fn intersects(self, other: Self) -> bool

Do the two ranges intersect?

assert!(Rangef::new(0.0, 10.0).intersects(Rangef::new(5.0, 15.0)));
assert!(Rangef::new(0.0, 10.0).intersects(Rangef::new(5.0, 6.0)));
assert!(Rangef::new(0.0, 10.0).intersects(Rangef::new(10.0, 20.0)));
assert!(!Rangef::new(0.0, 10.0).intersects(Rangef::new(20.0, 30.0)));