Struct emath::Vec2

pub struct Vec2 {
    pub x: f32,
    pub y: f32,
}

A vector has a direction and length. A Vec2 is often used to represent a size.

emath represents positions using crate::Pos2.

Normally the units are points (logical pixels).

Fields
x: f32y: f32
Implementations

impl Vec2

pub const X: Vec2

pub const Y: Vec2

pub const ZERO: Self

pub const INFINITY: Self

pub fn zero() -> Self

👎 Deprecated:
Use Vec2::ZERO instead

pub fn infinity() -> Self

👎 Deprecated:
Use Vec2::INFINITY instead

pub const fn new(x: f32, y: f32) -> Self

pub const fn splat(v: f32) -> Self

Set both x and y to the same value.

#[must_use]pub fn normalized(self) -> Self

Safe normalize: returns zero if input is zero.

pub fn rot90(self) -> Self

Rotates the vector by 90°, i.e positive X to positive Y (clockwise in egui coordinates).

pub fn length(self) -> f32

pub fn length_sq(self) -> f32

pub fn angled(angle: f32) -> Self

Create a unit vector with the given angle (in radians).

An angle of zero gives the unit X axis.

An angle of 𝞃/4 = 90° gives the unit Y axis.

#[must_use]pub fn floor(self) -> Self

#[must_use]pub fn round(self) -> Self

#[must_use]pub fn ceil(self) -> Self

pub fn is_finite(self) -> bool

True if all members are also finite.

pub fn any_nan(self) -> bool

True if any member is NaN.

#[must_use]pub fn min(self, other: Self) -> Self

#[must_use]pub fn max(self, other: Self) -> Self

#[must_use]pub fn min_elem(self) -> f32

Returns the minimum of self.x and self.y.

#[must_use]pub fn max_elem(self) -> f32

Returns the maximum of self.x and self.y.

#[must_use]pub fn clamp(self, min: Self, max: Self) -> Self